clients/finance: rewards pages in finance for maintainers
