+++ clients/apps/web/next.config.js
@@ -232,8 +232,8 @@ const nextConfig = {
         permanent: false,
       },
       {
-        source: '/maintainer/:organization/finance_new',
-        destination: '/maintainer/:organization/finance_new/incoming',
+        source: '/maintainer/:organization/finance',
+        destination: '/maintainer/:organization/finance/incoming',
         permanent: false,
       },
