+++ clients/apps/web/src/app/maintainer/[organization]/(topbar)/finance/incoming/ClientPage.tsx
@@ -28,9 +28,7 @@ export default function ClientPage() {
 
   const setActiveTab = useCallback(
     (value: string) => {
-      router.replace(
-        `/maintainer/${org?.name}/finance_new/incoming?type=${value}`,
-      )
+      router.replace(`/maintainer/${org?.name}/finance/incoming?type=${value}`)
     },
     [org],
   )
