+++ clients/apps/web/src/app/maintainer/[organization]/(topbar)/finance/issue-funding/ClientPage.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import Finance from '@/components/Finance/Finance'
-import Head from 'next/head'
-import { useRouter } from 'next/navigation'
+import { useToast } from '@/components/Toast/use-toast'
+import { useRouter, useSearchParams } from 'next/navigation'
 import {
   useAccount,
   useListPledgesForOrganization,
@@ -15,6 +15,11 @@ export default function ClientPage() {
   const router = useRouter()
   const { org, isLoaded } = useCurrentOrgAndRepoFromURL()
 
+  const params = useSearchParams()
+  const status = params?.get('status')
+
+  const { toast } = useToast()
+
   useEffect(() => {
     if (isLoaded && !org) {
       router.push('/feed')
@@ -22,21 +27,27 @@ export default function ClientPage() {
     }
   }, [isLoaded, org, router])
 
+  useEffect(() => {
+    if (status === 'stripe-connected') {
+      toast({
+        title: 'Stripe setup complete',
+        description: 'Your account is now ready to accept pledges.',
+      })
+    }
+  }, [status, toast])
+
   const pledges = useListPledgesForOrganization(org?.platform, org?.name)
   const rewards = useListRewards(org?.id)
   const { data: account } = useAccount(org?.account_id)
 
   return (
     <>
-      <Head>
-        <title>Polar{org ? ` ${org.name}` : ''}</title>
-      </Head>
       {org && pledges.data?.items && rewards.data?.items && (
         <Finance
           pledges={pledges.data.items}
           rewards={rewards.data.items}
           org={org}
-          tab="rewarded"
+          tab="current"
           account={account}
         />
       )}
