+++ clients/apps/web/src/app/maintainer/[organization]/(topbar)/finance/issue-funding/rewarded/ClientPage.tsx
@@ -1,24 +1,19 @@
 'use client'
 
 import Finance from '@/components/Finance/Finance'
-import { useToast } from '@/components/Toast/use-toast'
-import { useRouter, useSearchParams } from 'next/navigation'
+import Head from 'next/head'
+import { useRouter } from 'next/navigation'
 import {
   useAccount,
   useListPledgesForOrganization,
   useListRewards,
 } from 'polarkit/hooks'
 import { useEffect } from 'react'
-import { useCurrentOrgAndRepoFromURL } from '../../../../../hooks'
+import { useCurrentOrgAndRepoFromURL } from '../../../../../../../hooks'
 
 export default function ClientPage() {
   const router = useRouter()
   const { org, isLoaded } = useCurrentOrgAndRepoFromURL()
-
-  const params = useSearchParams()
-  const status = params?.get('status')
-
-  const { toast } = useToast()
 
   useEffect(() => {
     if (isLoaded && !org) {
@@ -27,27 +22,21 @@ export default function ClientPage() {
     }
   }, [isLoaded, org, router])
 
-  useEffect(() => {
-    if (status === 'stripe-connected') {
-      toast({
-        title: 'Stripe setup complete',
-        description: 'Your account is now ready to accept pledges.',
-      })
-    }
-  }, [status, toast])
-
   const pledges = useListPledgesForOrganization(org?.platform, org?.name)
   const rewards = useListRewards(org?.id)
   const { data: account } = useAccount(org?.account_id)
 
   return (
     <>
+      <Head>
+        <title>Polar{org ? ` ${org.name}` : ''}</title>
+      </Head>
       {org && pledges.data?.items && rewards.data?.items && (
         <Finance
           pledges={pledges.data.items}
           rewards={rewards.data.items}
           org={org}
-          tab="current"
+          tab="rewarded"
           account={account}
         />
       )}
